When fetching commits, stop if a tagged object is encountered
As a workaround for cases where the lower bound is a tag on a commit
of another branch, which we would never encounter. Effectively we'd
not stop iterating and include all commits made since the first ever
commit.

This happened on repositories with parallel releases being made from
multiple branches, losing linearity. That's not a use case I'm eager
to support, but we can at least reduce noise in our output.